chore(mock myinfo constants): shifts constants to msw/mocks
