+++ frontend/src/features/admin-form/create/builder-and-design/constants.ts
@@ -122,10 +122,3 @@ export const CREATE_MYINFO_PARTICULARS_DROP_ID = 'create-myinfo-particulars'
 export const CREATE_MYINFO_MARRIAGE_DROP_ID = 'create-myinfo-marriage'
 export const FIELD_LIST_DROP_ID = 'formFieldList'
 export const PENDING_CREATE_FIELD_ID = 'FIELD-PENDING-CREATION'
-
-export const MOCK_MYINFO_IMPLEMENTED_TYPES: MyInfoImplementedTypes[] = [
-  ...CREATE_MYINFO_PERSONAL_FIELDS_ORDERED,
-  ...CREATE_MYINFO_CONTACT_FIELDS_ORDERED,
-  ...CREATE_MYINFO_PARTICULARS_FIELDS_ORDERED,
-  ...CREATE_MYINFO_MARRIAGE_FIELDS_ORDERED,
-]
